requirements.txt
flask
pycryptodome
wfastcgi